requirements.txt
Django==2.0.7
djangorestframework==3.8.2
pandas==1.0.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
